# src/docs/swagger.yaml
basePath: /api/magic-circle
definitions:
...
  profile.GetAllProfilesResp:
    properties:
      profiles:
        items:
          $ref: '#/definitions/profile.Profile'
        type: array
    type: object
  profile.Profile:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  profile.UpdateProfileReq:
    properties:
      name:
        type: string
    type: object
  profile.UpdateProfileResp:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a server to get projects from github
  title: Magic-Circle API
  version: "1.0"
paths:
...
  /v1/profile:
    get:
      description: return all profiles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/profile.GetAllProfilesResp'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get all profiles
      tags:
      - profile
  /v1/profile/{id}:
    delete:
      description: Delete profile by id
      parameters:
      - description: id of profile
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete profile by id
      tags:
      - profile
    put:
      consumes:
      - application/json
      description: update profile
      parameters:
      - description: id of profile
        in: path
        name: id
        required: true
        type: integer
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/profile.UpdateProfileReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/profile.UpdateProfileResp'
        "400":
          description: ""
        "404":
          description: ""
        "500":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: Update profile
      tags:
      - profile
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"